+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 * Python: Added GETEX command ([#1612](https://github.com/aws/glide-for-redis/pull/1612))
 * Python: Added BITFIELD and BITFIELD_RO commands ([#1615](https://github.com/aws/glide-for-redis/pull/1615))
 * Python: Added ZREVRANK command ([#1614](https://github.com/aws/glide-for-redis/pull/1614))
+* Python: Added XDEL command ([#1619](https://github.com/aws/glide-for-redis/pull/1619))
 
 ### Breaking Changes
 * Node: Update XREAD to return a Map of Map ([#1494](https://github.com/aws/glide-for-redis/pull/1494))